# HealthPassProject/openapi.yaml
openapi: 3.0.3
info:
  title: HealthPass API
  description: API documentation for HealthPass application - a healthcare management platform
  version: 1.0.0
  contact:
    name: HealthPass Support
servers:
  - url: '{baseUrl}/api'
    description: API Server
    variables:
      baseUrl:
        default: http://localhost:5000
        description: Base URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /bills:
    get:
      tags:
        - Bills
      summary: Get bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - APPROVED
              - REJECTED
              - NEED_INFO
      responses:
        '200':
          description: Bills retrieved
    post:
      tags:
        - Bills
      summary: Upload bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                hospitalId:
                  type: string
                serviceDate:
                  type: string
                  format: date
                receiptUrl:
                  type: string
                  format: uri
                description:
                  type: string
      responses:
        '201':
          description: Bill uploaded
...
  /insurance:
    get:
      tags:
        - Insurance
      summary: Get insurances
      responses:
        '200':
          description: Insurances retrieved
    post:
      tags:
        - Insurance
      summary: Link insurance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - company
                - policyNumber
                - insuredName
                - expiryDate
              properties:
                company:
                  type: string
                policyNumber:
                  type: string
                insuredName:
                  type: string
                expiryDate:
                  type: string
                  format: date
                premiumDue:
                  type: string
                  format: date
      responses:
        '201':
          description: Insurance linked
  /insurance/{id}:
    delete:
      tags:
        - Insurance
      summary: Unlink insurance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Insurance unlinked
...
  /admin/bills/{id}/approve:
    patch:
      tags:
        - Admin
      summary: Approve bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - creditAmount
              properties:
                creditAmount:
                  type: number
                note:
                  type: string
      responses:
        '200':
          description: Bill approved
  /admin/bills/{id}/reject:
    patch:
      tags:
        - Admin
      summary: Reject bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Bill rejected
  /admin/bills/{id}/request-info:
    patch:
      tags:
        - Admin
      summary: Request more info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
      responses:
        '200':
          description: Info requested
  /admin/appointments/{id}/adjust-visits:
    patch:
      tags:
        - Admin
      summary: Adjust visits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - adjustment
                - reason
              properties:
                adjustment:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Visits adjusted
  /admin/audit-logs:
    get:
      tags:
        - Admin
      summary: Get audit logs
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Audit logs